src/requirements.txt
aiofiles
aiohttp
aiojobs
async-lru
async-property
click
colorlog
diskcache
fastapi==0.42.0
lazy-object-proxy
ldap3
PyJWT
pluggy
pydantic[ujson,email]==0.32.2
python-multipart
pytest-asyncio
requests
service
uvicorn
zmq
